Backend/requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.9.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
numpy==1.16.1
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.0
protobuf==3.9.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
scikit-learn==0.21.3
scipy==1.1.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.3
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
